Credential_Scanner-main/requirements.txt
fastapi
uvicorn
python-multipart
pymupdf
pillow
pytesseract
python-docx
pydantic
nltk
ollama  